double to float, to optimize for GPU

+++ JoinTest.h
@@ -13,18 +13,18 @@ public:
 	Grid *gridIndex;
 	vector<size_t> taskSet1, taskSet2;
 	int init(vector<STTraj> *dataPtr, Grid *gridIndex);
-	int defaultTest(double epsilon, double alpha, int setSize1, int setSize2, map<trajPair, double>& result);
+	int defaultTest(float epsilon, float alpha, int setSize1, int setSize2, map<trajPair, float>& result);
 	// join operations...
-	int joinExhaustedCPU(double epsilon,
-		double alpha,
+	int joinExhaustedCPU(float epsilon,
+		float alpha,
 		vector<size_t> &join_set_1,
 		vector<size_t> &join_set_2,
-		map<trajPair, double> &result);
-	int joinExhaustedGPU(double epsilon,
-		double alpha,
+		map<trajPair, float> &result);
+	int joinExhaustedGPU(float epsilon,
+		float alpha,
 		vector<size_t> &join_set_1,
 		vector<size_t> &join_set_2,
-		map<trajPair, double> &result);
+		map<trajPair, float> &result);
 
 	JoinTest();
 	~JoinTest();
